+++ arch/x86/rsdp.h
@@ -26,7 +26,7 @@ struct RSDPDescriptor
     uint8_t Checksum;
     char OEMID[6];
     uint8_t Revision;
-    uint32_t RsdtAddress;
+    uintptr_t RsdtAddress;
 } __attribute__((packed));
 
 struct RSDPDescriptor20
