+++ lib/libc/pthread.c
@@ -25,7 +25,7 @@ __no_return void
 pthread_exit(void *retptr)
 {
     __unused(retptr);
-    syscall(SYS_texit, (uint32_t) retptr, 0, 0);
+    syscall(SYS_texit, (uintptr_t) retptr, 0, 0);
     for (;;);
 }
